only move $T/environment if it exists; bug 64926
svn path=/main/trunk/; revision=3151

+++ bin/ebuild.sh
@@ -725,7 +725,7 @@ dyn_clean() {
 	if ! hasq keeptemp $FEATURES; then
 		rm -rf "${T}"
 	else
-		mv "${T}/environment" "${T}/environment.keeptemp"
+		[ -e "${T}/environment ] && mv "${T}/environment" "${T}/environment.keeptemp"
 	fi
 
 	if ! hasq keepwork $FEATURES; then
